- Attempted to fix a problem when Squid ignores all header fields it cannot   parse.  Now we support a two-level scheme where raw header fields are   preserved regardless of our parsing/interpretation abilities. Each header   field has also a "cache" where we store parsed value of the field (if any). - The only opearation on raw headers left is "joining" of appropriate headers   together. - Fixed a lot of minor problems and added new header fields:  Server, X-Cache,   and other.

+++ include/Stack.h
@@ -1,5 +1,5 @@
 /*
- * $Id: Stack.h,v 1.4 1998/03/03 00:30:55 rousskov Exp $
+ * $Id: Stack.h,v 1.5 1998/03/11 22:18:42 rousskov Exp $
  *
  * AUTHOR: Alex Rousskov
  *
@@ -37,7 +37,7 @@ struct _Stack {
     size_t capacity;
 
     /* protected, do not use these, use interface functions instead */
-    size_t count;
+    int count;
     void **items;
 };
